Documented the entire LVMiniAPI with Swagger and started a unit test project.

+++ LVMiniApi/Controllers/ProductGroupsController.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using Data.Service.Core.Entities;
 using Data.Service.Core.Interfaces;
+using LVMiniApi.Filters;
 using LVMiniApi.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,19 +11,33 @@ using System.Threading.Tasks;
 
 namespace LVMiniApi.Controllers
 {
+    /// <summary>
+    /// Controller for manipulating ProductGroups and the products in them.
+    /// </summary>
     [Route("api/productgroups")]
-    public class ProducGroupsController : BaseController
+    [Produces("application/json")]
+    [Authorize]
+    public class ProductGroupsController : BaseController
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly IProductGroupRepository _productGroupRepository;
 
-        public ProducGroupsController(IUnitOfWork unitOfWork, IMapper mapper)
+        /// <summary>
+        /// Injects the services needed through constructor injection.
+        /// </summary>
+        /// <param name="unitOfWork">Unit Of Work</param>
+        /// <param name="mapper">AutoMapper's Mapper class.</param>
+        public ProductGroupsController(IUnitOfWork unitOfWork, IMapper mapper)
         {
             _unitOfWork = unitOfWork;
             _productGroupRepository = _unitOfWork.ProductGroupRepository;
             Mapper = mapper;
         }
 
+        /// <summary>
+        /// Gets all existing ProductGroups in the database.
+        /// </summary>
+        /// <returns>Http 200 OK and a collection of ProductGroups.</returns>
         [HttpGet]
         public IActionResult GetAllProductGroups()
         {
@@ -31,6 +47,14 @@ namespace LVMiniApi.Controllers
             return Ok(productGroups);
         }
 
+        /// <summary>
+        /// Gets a specific ProductGroup based on its name.
+        /// </summary>
+        /// <param name="name">The name of the ProductGroup.</param>
+        /// <returns>
+        /// Http 200 OK and a single ProductGroup of the requested group.
+        /// Http 404 NotFound if there is no such group.
+        /// </returns>
         [HttpGet("{name}", Name = "ProductGroupGet")]
         public async Task<IActionResult> GetSingleProductGroup(string name)
         {
@@ -40,11 +64,21 @@ namespace LVMiniApi.Controllers
                 return NotFound();
             }
 
-            var prodcutGroupToReturn = Mapper.Map<ProductGroupDto>(productGroup);
-            return Ok(prodcutGroupToReturn);
+            var productGroupToReturn = Mapper.Map<ProductGroupDto>(productGroup);
+            return Ok(productGroupToReturn);
         }
 
+        /// <summary>
+        /// Creates a new ProductGroup with an optional initial collection of products. It could be empty as well.
+        /// </summary>
+        /// <param name="productGroup">The ProductGroup information needed to create a new one.</param>
+        /// <returns>
+        /// Http 201 Created with the created ProductGroup.
+        /// Http 400 BadRequest if the group already exists.
+        /// Http 500 if the insert to the database fails.
+        /// </returns>
         [HttpPost]
+        [ValidateModel]
         public async Task<IActionResult> AddProductGroup([FromBody] CreateProductGroupDto productGroup)
         {
             if (await _productGroupRepository.ProductGroupExists(productGroup.Name))
@@ -60,12 +94,20 @@ namespace LVMiniApi.Controllers
                 return new StatusCodeResult(500);
             }
 
-            var createdProductGroup = await _productGroupRepository.GetById(entity.IDProductGroup);
-            var groupToReturn = Mapper.Map<ProductGroupDto>(createdProductGroup);
+            var createdProductGroup = await _productGroupRepository.GetProductGroupByName(entity.Name);
 
-            return Ok(groupToReturn);
+            return CreatedAtRoute("ProductGroupGet", new { name = createdProductGroup.Name.ToLower() }, Mapper.Map<ProductGroupDto>(createdProductGroup));
         }
 
+        /// <summary>
+        /// Adds an existing product to an existing ProductGroup.
+        /// </summary>
+        /// <param name="name">The ProductGroup name.</param>
+        /// <param name="productCode">The product code.</param>
+        /// <returns>
+        /// Http 200 OK, the ProductGroup with an updated list of products.
+        /// Http 404 NotFound if either the ProductGroup or Product don't exist.
+        /// </returns>
         [HttpPost("{name}/products/{productCode}")]
         public async Task<IActionResult> AddProductToProductGroup(string name, string productCode)
         {
@@ -86,11 +128,20 @@ namespace LVMiniApi.Controllers
 
             var updatedProductGroup = await _productGroupRepository.GetById(productGroup.IDProductGroup);
             var groupToReturn = Mapper.Map<ProductGroupDto>(updatedProductGroup);
-
             return Ok(groupToReturn);
         }
 
+        /// <summary>
+        /// Updates an existing ProductGroup's information.
+        /// </summary>
+        /// <param name="name">The name of the group.</param>
+        /// <param name="productGroup">The fields you want ot update. All are optional.</param>
+        /// <returns>
+        /// Http 200 OK with the updated ProductGroup.
+        /// Http 404 NotFound if the group doesn't exist.
+        /// </returns>
         [HttpPatch("{name}")]
+        [ValidateModel]
         public async Task<IActionResult> UpdateGroup(string name, [FromBody] UpdateProductGroupDto productGroup)
         {
             var productGroupEntity = await _productGroupRepository.GetProductGroupByName(name);
@@ -106,9 +157,22 @@ namespace LVMiniApi.Controllers
             return Ok(modelToReturn);
         }
 
+        /// <summary>
+        /// Deletes an existing ProductGroup and the reference to it's products.
+        /// </summary>
+        /// <param name="name">The name of the group.</param>
+        /// <returns>
+        /// Http 204 NoContent if the delete was successful.
+        /// Http 404 NotFound if the group does not exist.
+        /// </returns>
         [HttpDelete("{name}")]
         public async Task<IActionResult> DeleteProductGroup(string name)
         {
+            if (!await _productGroupRepository.ProductGroupExists(name))
+            {
+                return NotFound();
+            }
+
             var productGroup = await _productGroupRepository.GetProductGroupByName(name);
 
             _productGroupRepository.Delete(productGroup);
@@ -117,6 +181,15 @@ namespace LVMiniApi.Controllers
             return NoContent();
         }
 
+        /// <summary>
+        /// Deletes a Product from a ProductGroup.
+        /// </summary>
+        /// <param name="name">The name of the group.</param>
+        /// <param name="productCode">The code of the Product.</param>
+        /// <returns>
+        /// Http 200 OK if the product was deleted and the group with an updated list of products.
+        /// Http 404 NotFound if either the ProductGroup or the Product don't exist.
+        /// </returns>
         [HttpDelete("{name}/products/{productCode}")]
         public async Task<IActionResult> RemoveProductFromGroup(string name, string productCode)
         {
@@ -136,6 +209,7 @@ namespace LVMiniApi.Controllers
             await _unitOfWork.Commit();
 
             var updatedProductGroup = await _productGroupRepository.GetById(productGroup.IDProductGroup);
+
             var groupToReturn = Mapper.Map<ProductGroupDto>(updatedProductGroup);
 
             return Ok(groupToReturn);
